Make sure the redirection is safe and cannot be abused
RH infosec team notified me that some people do
use incomplete redirection to construct specific URL when
the redirection do not include a trailling slash.

For example, a redirection from example.org to www.example.org
might be used as example.org/.example.com to
www.example.org.example.com, and a user or a filter would see the
redirection is incorrect. See a example on
http://git.theopensourceway.org/cgit/ansible/commit/?id=20b224a7d5cd8d4142b86ff657ffcd1a4fd30340