Fix some strange error regarding SSL
For a reason I didn't understood yet, the port comparaison became
invalid, because port is a integer and not a string. This
wasn't detected sooner, because httpd didn't restart nor crashed in
production, so the right configuration was still there.

I suspect a ansible change somewhere, but I am not sure when,
I might dig later.