Set ServerName to a invalid value
It was found while investigating #12 that _default_ vhost
sometime take over the vhost configuration we did define. The
condition are rather tricky to see something:
- we need a SSL enabled host
- it need to be on Apache 2.4
- the hostname must be the same as the vhost
- the hostname must be after ssl.conf using the current
file loading order (currently based on filename and alphabetical
order).

So the trick is either to change the load order (which is IMHO
too much hassle), or override the default Servername to avoid
problem.

Should fix #12