Use default rather than set_fact for _website_domain
Main reason is that the munin role is using some trick to
be deployed on the whole server for the client, and just
on the server for the server part. However, this trick break now
since _website_domain is undefined (since that's done dynamically
with set_fact, who is skipped).

So I have to make sure there is a deault value in vars to make it
run.