Make letsencrypt work also with reverse proxy
That's basically the same kind of issues fixed by 4e22d8b42.
The proxy should move everything but the letsencrypt negociation.

This part is also needing a bit of refactoring anyway, but this
is the least intrusive fix I could find for now.

Using ProxyPassmatch didn't seems to work on 2.2, so I have to use mod_rewrite